auth module moved to api/v1, fixed google callback, fixed front-end query user
