+++ backend/api/v1/index.js
@@ -2,7 +2,7 @@ const express = require('express');
 
 module.exports = (cache) => {
     const router = express.Router();
-
+    router.use('/auth', require('./auth')());
     router.use('/blog', require('./blog')(cache));
     router.use('/contributors', require('./contributors')(cache));
     router.use('/episodes', require('./episodes')(cache));
